zhijian_langgraph/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymysql==1.1.1
PyJWT==2.10.1
httpx==0.28.1
redis==5.2.1
langgraph>=1.0.0,<2
langgraph-checkpoint>=3.0.1,<5
